# public/openapi.spec.json
{
  "components": {
    "schemas": {
...
      "SalonRequest": {
        "type": "object",
        "required": [
          "companyId",
          "locationId",
          "description",
          "name",
          "temporalDurationId"
        ],
        "properties": {
          "companyId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "temporalDurationId": {
            "type": "string",
            "description": "Default time slot duration for the salon"
          }
        }
      },
      "SalonUpdateRequest": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "temporalDurationId": {
            "type": "string"
          }
        }
      },
      "Salon": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "companyId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "temporalDurationId": {
            "type": "string"
          },
          "seatIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "staffIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "salonOpenHourIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Staff": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "staffRoleId": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          },
          "forename": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "haircutIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "StaffUpdateRequest": {
        "type": "object",
        "properties": {
          "staffRoleId": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "forename": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          }
        }
      },
      "TimeSlotAvailability": {
        "type": "object",
        "properties": {
          "salonTimeSlotSequenceNumber": {
            "type": "integer"
          },
          "startTime": {
            "type": "string"
          },
          "endTime": {
            "type": "string"
          },
          "available": {
            "type": "boolean"
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - Invalid input parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
  "info": {
    "title": "Nova API",
    "version": "2.0.0",
    "description": "Nova Backend API - A booking management system for workplace/salon bookings with support for services, staff, and appointments"
  },
  "openapi": "3.0.4",
  "paths": {
...
    "/v2/salon": {
      "post": {
        "summary": "Create a salon",
        "description": "Creates a new salon/workplace location",
        "tags": ["Salon"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalonRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Salon created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Salon"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/v2/salon/{salonId}": {
      "patch": {
        "summary": "Update a salon",
        "description": "Updates an existing salon's properties",
        "tags": ["Salon"],
        "parameters": [
          {
            "name": "salonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The salon ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SalonUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Salon updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Salon"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v2/salon/{salonId}/product/{serviceId}/staff/{staffId}": {
      "get": {
        "summary": "Get staff availability for service",
        "description": "Retrieves available time slots for a specific staff member to perform a service",
        "tags": ["Salon"],
        "parameters": [
          {
            "name": "salonId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The salon ID"
          },
          {
            "name": "serviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The service/product ID"
          },
          {
            "name": "staffId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The staff member ID"
          },
          {
            "name": "startDateYear",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Year of the date to check availability"
          },
          {
            "name": "startDateMonth",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 12
            },
            "description": "Month of the date to check availability"
          },
          {
            "name": "startDateMonthDay",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 31
            },
            "description": "Day of the month to check availability"
          }
        ],
        "responses": {
          "200": {
            "description": "List of available time slots",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TimeSlotAvailability"
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/v2/staff/{staffId}": {
      "patch": {
        "summary": "Update staff member",
        "description": "Updates a staff member's details",
        "tags": ["Staff"],
        "parameters": [
          {
            "name": "staffId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The staff ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Staff updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Staff"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
}